feat: rewrite groundwork plugin using the gwos/tcg library to make the code shorter
