+++ plugins/outputs/all/all.go
@@ -21,7 +21,7 @@ import (
 	_ "github.com/influxdata/telegraf/plugins/outputs/file"
 	_ "github.com/influxdata/telegraf/plugins/outputs/graphite"
 	_ "github.com/influxdata/telegraf/plugins/outputs/graylog"
-	_ "github.com/influxdata/telegraf/plugins/outputs/gw8"
+	_ "github.com/influxdata/telegraf/plugins/outputs/groundwork"
 	_ "github.com/influxdata/telegraf/plugins/outputs/health"
 	_ "github.com/influxdata/telegraf/plugins/outputs/http"
 	_ "github.com/influxdata/telegraf/plugins/outputs/influxdb"
