+++ plugins/outputs/groundwork/README.md
@@ -1,21 +1,18 @@
 # Groundwork Output Plugin
 
-This plugin writes to a Groundwork instance using "[GROUNDWORK][]" own format.
+This plugin writes to a [GroundWork Monitor][1] instance.
 
-[GROUNDWORK]: https://www.gwos.com
+[1]: https://www.gwos.com/product/groundwork-monitor/
 
 ### Configuration:
 
 ```toml
-[[outputs.gw8]]
+[[outputs.groundwork]]
   ## HTTP endpoint for your groundwork instance.
-  endpoint = ""
+  groundwork_endpoint = ""
 
   ## Agent uuid for Groundwork API Server
   agent_id = ""
-
-  ## Groundwork application type
-  app_type = ""
 
   ## Username to access Groundwork API
   username = ""
@@ -23,8 +20,8 @@ This plugin writes to a Groundwork instance using "[GROUNDWORK][]" own format.
   ## Password to use in pair with username
   password = ""
   
-  ## Default display name for the host with services(metrics)
-  default_host = "default_telegraf"
+  ## Default display name for the host with services(metrics) [default - "telegraf"]
+  default_host = "telegraf"
 
   ## Default service state [default - "SERVICE_OK"]
   default_service_state = "SERVICE_OK"
@@ -32,3 +29,4 @@ This plugin writes to a Groundwork instance using "[GROUNDWORK][]" own format.
   ## The name of the tag that contains the hostname [default - "host"]
   resource_tag = "host"
 ```
+#### NOTE: Plugin only supports GW8+
